Update namespaces in files
